+++ src/Phan/Language/Internal/FunctionSignatureMap.php
@@ -45,8 +45,18 @@ PHAN;
  * 2. '&w_name' indicates that a parameter is expected to be passed in, and the value will be ignored, and may be overwritten.
  *
  * This file contains the signatures for the most recent minor release of PHP supported by phan (php 7.2)
+ *
+ * Changes:
+ *
+ * In Phan 0.12.3,
+ *
+ * - This started using array shapes for union types (array{...}).
+ *
+ *   \Phan\Language\UnionType->withFlattenedArrayShapeTypeInstances() may be of help to programatically convert these to array<string,T1>|array<string,T2>
+ *
+ * - This started using array shapes with optional fields for union types (array{key:int=}).
+ *   The `=` after the union type indicates that the field is optional.
  */
-
 return [
 '__halt_compiler' => [''],
 'abs' => ['int', 'number'=>'int'],
@@ -276,6 +286,7 @@ return [
 'ast\get_metadata' => ['array<int,ast\Metadata>'],
 'ast\get_supported_versions' => ['array<int,int>', 'exclude_deprecated='=>'bool'],
 'ast\kind_uses_flags' => ['bool', 'kind'=>'int'],
+'ast\Node::__construct' => ['void', 'kind='=>'int', 'flags='=> 'int', 'children='=>'ast\Node\Decl[]|ast\Node[]|array[]|int[]|string[]|float[]|bool[]|null[]', 'start_line='=>'int'],
 'ast\parse_code' => ['ast\Node', 'code'=>'string', 'version'=>'int', 'filename='=>'string'],
 'ast\parse_file' => ['ast\Node', 'filename'=>'string', 'version'=>'int'],
 'atan' => ['float', 'number'=>'float'],
@@ -6699,7 +6710,7 @@ return [
 'parse_ini_file' => ['array|false', 'filename'=>'string', 'process_sections='=>'bool', 'scanner_mode='=>'int'],
 'parse_ini_string' => ['array|false', 'ini_string'=>'string', 'process_sections='=>'bool', 'scanner_mode='=>'int'],
 'parse_str' => ['void', 'encoded_string'=>'string', '&w_result='=>'array'],
-'parse_url' => ['mixed', 'url'=>'string', 'url_component='=>'int'],
+'parse_url' => ['array{scheme:string=,host:string=,port:int=,user:string=,pass:string=,path:string=,query:string=,fragment:string=}|string|int|null', 'url'=>'string', 'url_component='=>'int'],
 'ParseError::__clone' => ['void'],
 'ParseError::__construct' => ['void', 'message='=>'string', 'code='=>'int', 'previous='=>'?Throwable|?ParseError'],
 'ParseError::__toString' => ['string'],
@@ -10211,7 +10222,7 @@ return [
 'unlink' => ['bool', 'filename'=>'string', 'context='=>'resource'],
 'unpack' => ['array', 'format'=>'string', 'data'=>'string', 'offset='=>'int'],
 'unregister_tick_function' => ['void', 'function_name'=>'string'],
-'unserialize' => ['mixed', 'variable_representation'=>'string', 'allowed_classes='=>'array'],
+'unserialize' => ['mixed', 'variable_representation'=>'string', 'allowed_classes='=>'array{allowed_classes:string[]|bool=}'],
 'unset' => ['void', 'var='=>'mixed', '...='=>'mixed'],
 'untaint' => ['bool', '&rw_string'=>'string', '&...rw_strings='=>'string'],
 'uopz_allow_exit' => ['void', 'allow'=>'bool'],