+++ tests/files/src/0445_unserialize_allowed_classes.php
@@ -1,0 +1,6 @@
+<?php
+
+$serialized = serialize(new stdClass());
+unserialize($serialized, []);
+unserialize($serialized, ['allowed_classes' => [['stdClass']]]);
+unserialize($serialized, ['allowed_classes' => ['stdClass']]);