+++ tests/files/src/0446_new_node.php
@@ -1,0 +1,8 @@
+<?php
+var_export(new ast\Node());
+var_export(new ast\Node(
+    'key',
+    2,
+    [new stdClass()],
+    4
+));